Modified the calculation count histogram to use the kmer_DB class and changed the name of the program from kmer_hist to YV_kmerhist from now I will try to change all the usable programs to be of the name YV_
I also commented out all the previous "main" of kmer_DB, which loaded the kmers that apeared more than once and went over all the DB files.
I will add this logic in a new YV_ file.

Moreover, kmer_general.h had definition of function which created problems if two cpp files included the h file (as it is represented as two equal symbols)
I changed the function to be inline but this is just a patch and should be organized better...