I moved the initial scoring for the k-mers to use a function montonic with GRAMMER-GEMMA score. This worked beutifully on FT10, which has become my bench-mark phenotype. FT10 is a good phenotype to evaluate the method, as it is both polygenic, it is extremley affected by popultation structure and the associations are very close to the threshold.
Most of the time from the last commit was directed to create a permutation test. Altough not clear from
the Arabidopsis GWAS literature, permutation test doesn't actually mean reordering the phenotype relative
to the genotype. One should randomize giving the covarience matrix of phenotypes, which is dependent
on hertability and the kinship between individuals. Implementing it turned out to be senstive on how it
is done. Now I have to procedures that can permute phenotypes giving the covarience.

Now that I have the permutation test and a scoring for k-mers which takes population structure into
account I will try to improve the code and make it more general. The main changes I want to make are:
1. Merge the K-mers DB once into a file, and then load this file
2. Support all k-mers of size <32