Change all the phenotypes to work with floats (from the computational tricks with uint64_t) Also change the R script not to convert phenotypes to uint
This was done as I saw that the tricks with unsigned ints did not give significant boost to speed.
This is true given the other speedups implemented recently.

Interestingly altough there is significant correlation between the scores with floats and with unsigned int,
the correlation is not perfect. Likley due to inprecision in floats/doubles.

Next step will be to incorperate implementation of the dot-product inside the scoring using
the special architecture of the proccessors. That is SSE4, and maybe also AVX2, though it works only
on part of the cluster and for now I don't want to be that specific.