Removed the dependency on the original GRAMMAR-Gamma code, as the only calculation we are use is multiplaying by the inverse covarianve matrix.
Note: I tried before to calculate the Gamma using the GenABEL code, I got wrong results. This was due to an assumption in GRAMMAR-Gamma on the method of calculating the relatedness matrix.
I add a calculation of the relatedness matrix using their method in the previous commit. Using this matrix I can calculate easily the Gamma constant.