Added the option of calculating Gamma on part of the kmers table during the association of k-mers. The program can now get as input a inverse of the covariance matrix and calculate Gamma using the presence absence patterns. For now we take only 1,000,000 k-mers for gamma calculation, as this is computationaly very intensive. taking only park of the k-mers is both reasonable given the theory of gamma calculation and also gave the exact same Gamma in practice for FT10 where I also calculated it exactly using the full data.
As even 1,000,000 k-mers takes some time, I run it on a different thread during the full progress of associations and outputting.
To be able to plot the QQ plot altought we only have the Gamma after association, the QQ plot class was modified to be given the
Gamma only upon outputting the graph information.